backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
cloudinary==1.44.0
Django==5.2
django-autoslug==1.9.9
django-cors-headers==4.7.0
django-simple-history==3.8.0
djangorestframework==3.16.0
drf-yasg==1.21.10
ecdsa==0.19.1
idna==3.10
inflection==0.5.1
mysqlclient==2.2.7
packaging==24.2
pillow==11.2.1
pyasn1==0.4.8
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.1
requests==2.32.4
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
